Extract `button.ts`
